requirements.txt
asgiref==3.8.1
Django==5.0.4
environs==11.0.0
factory-boy==3.3.0
Faker==25.8.0
iniconfig==2.0.0
marshmallow==3.21.1
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.1.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
